Move introspection query to views level
