+++ src/custom-apps/utils.tsx
@@ -1,3 +1,4 @@
+import { gql } from "@apollo/client";
 import { WebhookEventTypeAsyncEnum, WebhookFragment } from "@dashboard/graphql";
 
 export function isUnnamed(webhook: WebhookFragment | undefined): boolean {
@@ -15,3 +16,31 @@ export const filterSelectedAsyncEvents = (
   }
   return asyncEvents;
 };
+
+export interface IntrospectionNode {
+  name: string;
+  interfaces: Array<{
+    name: string;
+  }> | null;
+  description: string;
+}
+
+// cannot be in `queries.ts` as codegen cannot handle `__schema`
+export const IntrospectionQuery = gql`
+  query EventsIntrospection {
+    __schema {
+      types {
+        name
+        interfaces {
+          name
+        }
+        description
+      }
+    }
+  }
+`;
+
+const isEvent = ({ name }: { name: string }) => name === "Event";
+
+export const buildEventsMap = (elements: IntrospectionNode[]) =>
+  elements.filter(({ interfaces }) => (interfaces || []).some(isEvent));
